third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-allow-lists-starting-with-comma-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Allow lists starting with a comma should be parsed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Sending GET request with custom headers.
PASS xhr.send(null) is undefined.
PASS response['x-print'] is "unicorn"
PASS response['y-print'] is "narwhal"
Sending PUT request.
PASS xhr.send(null) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
